revert:
75d5b63 "if a Noteskin fails to be loaded by the time the notes need to be displayed, attempt to force the default theme on and throw errors in\
to the log rather than crashing SM."
There was no indication in the commit of when this was actually happening; this
is unnecessary logic that will do nothing but hide bugs.

c9c74c7, fd5f5d3
"Alternative Column Input Metrics -- Mainly a requirement for Ez2dancer gametype. Allows for a themer to make game inputs that aren't usually used to trigger a column press do so in addition
to the usual buttons. (In the case of ez2dancer, the lower hand sensors are used as well as upper hand sensors, except for real mode where they are defined seperately)."
This is a gross hack.  If you want to write ugly hacks in other people's code
that badly, you can fork.

a7b2e4e "New Preference: OnlyPreferredDifficulties -- Drops songs that don't have steps of the preferred difficulty for the primary player. Useful if you want to simulate games that lock you
to a difficulty after a style/mode selection."
We don't need yet more code to emulate the *limitations* of other games.  This
codebase is way too bloated already.